daa: address Copilot review findings on Phase 2
PR #52 review pass. Four findings, all valid and addressed:

1. MSVC __int128 portability (shekyl_lwma1_hybrid_reference.h:67-68).
   The hybrid reference uses GCC/Clang `__int128` extension, which
   MSVC does not support. Gated the `lwma1-cross-check` CMake target
   behind `if(NOT MSVC)` so MSVC builds skip the harness cleanly
   instead of failing to compile. Documented the GCC/Clang-only
   requirement in the header. The reference body itself is left
   byte-identical to the vendored source per the Round 5 pinning
   discipline; the disposition is build-system-side, not code-side.
   Coverage of the underlying algorithm on MSVC remains via
   `cargo test -p shekyl-difficulty`, which is platform-independent.

2. printf-format consistency (lwma1_cross_check.cpp:95,100).
   Mixed `%zu` + `ts.size()` in two diagnostic sites with the
   `%llu` + `static_cast<unsigned long long>(...)` pattern used
   elsewhere in the same file. Normalized to the latter for
   stylistic consistency.

3. Markdown rendering of LWMA-3 verbatim extraction
   (zawy12_issue_3_lwma3.md:36-81). The verbatim C++ body was not
   wrapped in a code fence, so Markdown renderers parsed
   `std::vector<uint64_t>` as an HTML tag. Wrapped the body in a
   ```cpp fence; updated the file header to document that the fence
   delimiters are outside the byte-offset slice
   (`lwma3_byte_offset_start=17899`, `lwma3_byte_offset_end=20097`)
   and the fenced body remains byte-identical to the upstream slice
   (SHA-256 verified: both = 723bb80ce7d97d55f523783ea1904a72b915993883c9a42ede5d478099f27445).

4. Per-call `Vec<u128>` allocation in the FFI shim
   (difficulty_ffi.rs:194). Phase 2 disposition is "document, defer
   to V3.1 perf pass" per `15-deletion-and-debt.mdc` scope
   discipline. Reinterpret-casting `&[ShekylU128]` (align 8) to
   `&[u128]` (align 16) is exactly the layout coupling Round 5
   rejected when it chose the `ShekylU128` ABI; the materialized
   widening is unavoidable under the current contract. Added a
   "Performance / allocation" section to the module docs naming
   the ~1.5 KiB cost (`91 * 16` bytes) and the Phase 4 amortization
   posture. Added a V3.1 FOLLOWUPS entry listing three disposition
   options to compare against benchmark data at the V3.1 perf pass:
   (a) algorithm-signature change to iterator/const-generic buffer,
   (b) `heapless::Vec<u128, 91>` in the shim, (c) parallel
   `lwma1_next_decomposed(&[u64], &[(u64, u64)])` entry. The
   disposition is correctness-neutral (algorithm output is
   byte-identical regardless); only throughput changes.

Gates:
- cargo fmt --all -- --check: clean.
- cargo clippy -p shekyl-ffi -p shekyl-difficulty --all-targets
  -- -D warnings: clean.
- cargo test -p shekyl-ffi -p shekyl-difficulty: all pass.
- cargo doc -p shekyl-difficulty with
  -D rustdoc::broken_intra_doc_links
  -D rustdoc::private_intra_doc_links: clean.
- cmake --build . --target lwma1-cross-check: clean.
- ctest -R lwma1_cross_check: 1/1 pass.

Refs:
- PR #52 review thread (4 findings).
- `docs/design/DAA_LWMA1_PLAN.md` Phase 2 scope.
- `15-deletion-and-debt.mdc` (scope-per-PR; FOLLOWUPS target
  version requirement).

+++ tests/difficulty/lwma1_cross_check.cpp
@@ -92,14 +92,16 @@ uint64_t call_rust(const std::vector<uint64_t>& ts) {
     if (rc != SHEKYL_DIFFICULTY_OK) {
         std::fprintf(stderr,
             "FATAL: shekyl_difficulty_lwma1_next returned %d for input "
-            "of size %zu\n", rc, ts.size());
+            "of size %llu\n",
+            rc, static_cast<unsigned long long>(ts.size()));
         std::exit(2);
     }
     if (out.hi != 0) {
         std::fprintf(stderr,
-            "FATAL: Rust returned out.hi = 0x%llx for input of size %zu; "
+            "FATAL: Rust returned out.hi = 0x%llx for input of size %llu; "
             "no §8.1 vector should overflow u64\n",
-            static_cast<unsigned long long>(out.hi), ts.size());
+            static_cast<unsigned long long>(out.hi),
+            static_cast<unsigned long long>(ts.size()));
         std::exit(2);
     }
     return out.lo;
